deployment: libvirt: update machine type/model for all-in-one
Refer also to commit 7abcd94a, which applies the change to
controller.xml

Tests should run on a comparably newer model of CPU and should be fixed
for all tests under virtualization.

Nehalem cpu model is recommended as one that has been used for this
purpose for a few years.

Specifying 'forbid' for fallback and 'exact' for mode will ensure that
hosts the Qemu emulator is up-to-date on the host, capable of emulating
the expect model.  The consistency will make test results reliable
across host machines.

The newer machine type supports ACPI for example.

The feature "vmx" is required on computes to support nesting.

Related-Bug: #1793214

Change-Id: Ib3e72d8c9fc528978358a7054b337672284f2908
Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>