Fix check for ISO file in deployment script.
The checking for ISO files was made through the mime type. However
this method can fail as the mime type can be set differently by the
browser or tool used to download the file. An example of this
behavior can be found on Fedora 28 where the file command returns
application/octect-stream.

The fix is to avoid using the mime type for checking and just verify
that if a DOS/MBR file.

Closes-Bug: 1804610
Change-Id: I69f09df66215d405720801fbcbe3d43aac668564
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>